homework5 done
