+++ server/routes/index.js
@@ -4,7 +4,6 @@ const router = new Router();
 const handleRouter = require('./routes');
 const handleStreamRouter = require('./routesStream');
 
-// router.use('/store/csv', handleStreamRouter);
 router.use('/store', handleStreamRouter);
 router.use('/', handleRouter);
